requirements.txt
torch==1.3.1
torchvision==0.4.2
matplotlib==3.1.1
wandb==0.8.28
python-box==4.2.1